% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plot.simplex.noah.2}
\alias{plot.simplex.noah.2}
\title{plot.simplex.noah2}
\usage{
\method{plot}{simplex.noah.2}(theta, data, a = 1, x3 = 0.5, y3.off = 0,
  new.plot = T, xlm = c(0, 1), ylm = c(0, 1), pch = 19, alpha = 0.8,
  cex = 1.25, lex = 0.85, tscl = 1, labs = c("(1,0,0)", "(0,1,0)",
  "(0,0,1)"))
}
\arguments{
\item{a}{}

\item{x}{a vector of values between 0 and 1, representing the weight on the first dimension.}

\item{y}{a vector of values between 0 and 1, representing the weight on the second dimension.}

\item{z}{a vector of values between 0 and 1, representing the weight on the third dimension.}
}
\value{
plots a pretty graph
}
\description{
This function is by Noah Silbert
designed to prettily plot on the surfaces of simplexes.
It does not allow you to set colors
}
