% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plot.simplex.noah}
\alias{plot.simplex.noah}
\title{plot.simplex.noah}
\usage{
\method{plot}{simplex.noah}(theta, data, a = 1, x3 = 0.5, y3.off = 0,
  new.plot = T, xlm = c(0, 1), ylm = c(0, 1), xlab = "(1,0,0)",
  ylab = "(0,1,0)", zlab = "(0,0,1)")
}
\arguments{
\item{a}{}

\item{x}{a vector of values between 0 and 1, representing the weight on the first dimension.}

\item{y}{a vector of values between 0 and 1, representing the weight on the second dimension.}

\item{z}{a vector of values between 0 and 1, representing the weight on the third dimension.}
}
\value{
plots a pretty graph
}
\description{
This function is by Noah Silbert
designed to prettily plot on the surfaces of simplexes.
It does not allow you to set colors
}
