% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plot.simplex}
\alias{plot.simplex}
\title{plot.simplex}
\usage{
\method{plot}{simplex}(x, y, z, color = 0, a = 1, x3 = 0.5, y3.off = 0,
  new.plot = T, xlim = c(0, 1), ylim = c(0, 1), xlab = "Center",
  ylab = "Simple", zlab = "Dual", colorOption = "color")
}
\arguments{
\item{x}{a vector of values between 0 and 1, representing the weight on the first dimension.}

\item{y}{a vector of values between 0 and 1, representing the weight on the second dimension.}

\item{z}{a vector of values between 0 and 1, representing the weight on the third dimension.}

\item{color}{the color to scale everything in terms of}

\item{a}{}

\item{colorOption}{legal values include "color" and "grayscale"}
}
\value{
a pretty graph
}
\description{
This function adapts code by Noah Silbert, 
designed to prettily plot on the surfaces of simplexes.
It allows you to set colors
}
